Extrac schema into different file
